Keep manual replies instant (for integration test speeds)
Its sad that you can loose replies. But being fast is also cool if there
is no spam protection. Anyways the integration tests are more important
and they should be as fast as possible. So this would have to be fixed
smarter. Or at least in a way where integration tests still can go fast.