ux: refactor warlist color just to please clang
clang tidy started to detect some ``pName`` being null after the last
merge with ddnet.

And it is the one line of SetNameplateColor diff that chillerbot-ux
added which trips up clang.

So instead of taking a pointer which I do not know why it was the case
in the first place I now return a value.

It is a vec4 so returning 4 floats by value seems light weight enough to
do in a on render call.