ux: adapt reply to where test to new ddnet spawn physics
ddnet changed the spawning physics in

https://github.com/ddnet/ddnet/pull/10985

The test uses the default map Tutorial
which is a solo map. It used to be the case that the second tee
spawns on the left of the first tee.
Now both tees spawn at the exact same position.
Which the reply bot has no specific case for so it just
says "above" instead of "left"

Thanks kebso @KebsCS for breaking ma tests :angry: